Adding stations to map fragment

+++ app/src/main/java/com/respirar/model/Coordinates.kt
@@ -1,6 +1,6 @@
 package com.respirar.model
 
-class Coordinates(latitude: Float, longitude: Float) {
+class Coordinates(latitude: Double, longitude: Double) {
     val latitude = latitude
     val longitude = longitude
 }
